backend/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.2
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
Werkzeug==3.1.5
psycopg2-binary==2.9.9
